backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.7
sse-starlette==1.6.5
starlette>=0.27.0

# Database
sqlalchemy==2.0.27
psycopg2-binary==2.9.9

# Redis
redis[hiredis]>=4.2.0

# Google API
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-auth>=2.22.0
google-api-core>=2.11.0
protobuf>=4.21.0
googleapis-common-protos>=1.56.0

# Calendar parsing
icalendar==5.0.11

# Async support
aiofiles==23.2.1
anyio>=3.0.0

# Type hints
pydantic==2.6.1

# Type checking
types-redis>=4.2.0
types-requests>=2.31.0
mypy>=1.8.0 